🚚 rename 'testSetup' to 'framework'
