+++ test/e2e/lib/framework/eventsRegistry.ts
